chore(tests): move tests
