+++ api.rest
@@ -8,7 +8,7 @@ Content-Type: application/x-www-form-urlencoded
 email={{email}}&password={{password}}
 
 ###
-@authToken = {{login.response.body.payload.token}}
+@authToken = {{login.response.body.payload.jwt}}
 
 ### List All Bots (need to login first)
 GET {{baseUrl}}/api/v1/admin/bots
@@ -19,7 +19,7 @@ X-BP-Workspace: default
 @botId = YOUR_BOT_ID
 
 ### Send a request using Converse
-POST {{baseUrl}}/api/v1/bots/{{botId}}/converse/benchmarkUser
+POST {{baseUrl}}/api/v1/bots/welcome-bot/converse/benchmarkUser
 Content-Type: application/json
 
 {
