+++ config/jest.e2e.config.js
@@ -2,12 +2,12 @@ module.exports = {
   preset: 'jest-puppeteer',
   globals: {
     'ts-jest': {
-      tsConfig: '<rootDir>/src/tsconfig.test.json',
+      tsConfig: '<rootDir>/src/tests/tsconfig.test.json',
       isolatedModules: true
     }
   },
   globalSetup: 'jest-environment-puppeteer/setup',
-  setupFilesAfterEnv: ['expect-puppeteer', './jest.setup.js'],
+  setupFilesAfterEnv: ['expect-puppeteer', './config/jest.setup.js'],
   collectCoverage: false,
   resetModules: true,
   modulePaths: ['<rootDir>/src/bp/'],
@@ -15,10 +15,10 @@ module.exports = {
   transform: {
     '^.+\\.(ts|tsx|js)$': 'ts-jest'
   },
-  testMatch: ['**/src/e2e/**/*.test.(ts|js)'],
+  testMatch: ['**/src/tests/e2e/**/*.test.(ts|js)'],
   testPathIgnorePatterns: ['out', 'build', 'node_modules', 'src/bp'],
-  testEnvironment: '<rootDir>/src/e2e/customEnvironment.js',
-  rootDir: '.',
+  testEnvironment: '<rootDir>/src/tests/e2e/customEnvironment.js',
+  rootDir: '../',
   testResultsProcessor: './node_modules/jest-html-reporter',
-  testRunner: "jest-circus/runner",
+  testRunner: 'jest-circus/runner'
 }
