+++ src/tests/e2e/admin/logout.test.ts
@@ -1,6 +1,6 @@
 import axios from 'axios'
 
-import { bpConfig } from '../../../jest-puppeteer.config'
+import { bpConfig } from '../config'
 import { clickOn, expectMatchElement } from '../expectPuppeteer'
 import { CONFIRM_DIALOG, expectAdminApiCallSuccess, getResponse, gotoAndExpect } from '../utils'
