+++ src/tests/e2e/assets/alien.png
