+++ src/tests/e2e/assets/bot-import-test.tgz
