+++ src/tests/e2e/config.ts
@@ -1,0 +1,3 @@
+import { launch, bpConfig } from '../../../config/jest-puppeteer.config'
+
+export { launch, bpConfig }
