+++ src/tests/e2e/customEnvironment.js
@@ -5,7 +5,7 @@ class CustomEnvironment extends PuppeteerEnvironment {
     if (event.name == 'test_fn_failure') {
       const testName = state.currentlyRunningTest.name
       // Take a screenshot at the point of failure
-      await this.global.page.screenshot({ path: `build/tests/e2e/screenshots/${testName}.png` })
+      await this.global.page.screenshot({ path: `../build/tests/e2e/screenshots/${testName}.png` })
     }
   }
 }
