+++ src/tests/e2e/modules/code-editor.test.ts
