+++ src/tests/jest-rewire.ts
@@ -1,5 +1,5 @@
 import './jest-before'
-import rewire from './sdk/rewire'
+import rewire from '../bp/sdk/rewire'
 
 export = async () => {
   global.rewire = rewire
