+++ src/tsconfig.json
@@ -17,7 +17,7 @@
     "types": ["reflect-metadata", "jest"]
   },
   "exclude": [
-    "e2e/**",
+    "tests/e2e/**",
     "**/*.__test__.(ts|js)",
     "../out",
     "**/node_modules/**/*.(ts|js)",
